+++ tests/Test.hs
@@ -1,23 +1,28 @@
-import Control.Exception (catch)
-import System.Directory (removeDirectoryRecursive)
+module Main where
+
+import System.Directory (canonicalizePath, removePathForcibly)
 import System.FilePath (combine)
 import Test.Contrib
 import Test.Example
-import Test.Helpers (ioErrorHandler, rootPath)
 import Test.Shikensu
 import Test.Tasty
 import Test.Utilities
 
 
 main :: IO ()
-main =
-    rootPath
-        >>= \r -> catch
-            (removeDirectoryRecursive (combine r "tests/build"))
-            (ioErrorHandler)
-        >> defaultMain tests
+main = do
+    root        <- canonicalizePath "./"
+    _           <- removePathForcibly (combine root "tests/build")
+
+    defaultMain tests
 
 
 tests :: TestTree
 tests =
-    testGroup "Tests" [shikensuTests, contribTests, exampleTests, utilityTests]
+    testGroup
+        "Tests"
+        [ shikensuTests
+        , contribTests
+        , exampleTests
+        , utilityTests
+        ]