+++ tests/Test/Contrib.hs
@@ -20,7 +20,6 @@ import qualified Data.Text.IO as Text (readFile)
 import qualified Shikensu
 import qualified Shikensu.Contrib as Contrib
 import qualified Shikensu.Contrib.IO as Contrib.IO
-import qualified Shikensu.Types as Shikensu
 
 
 contribTests :: TestTree
@@ -45,7 +44,7 @@ contribTests = testGroup
 
 list :: String -> IO Shikensu.Dictionary
 list pattern =
-    rootPath >>= Shikensu.list [pattern]
+    Shikensu.listRelative [pattern] "."
 
 
 example_md :: IO Shikensu.Dictionary
@@ -72,10 +71,10 @@ testClone :: TestTree
 testClone =
     let
         dictionary = fmap (Contrib.clone "example.md" "cloned.md") example_md
-        definition = fmap (List.head . List.reverse) dictionary
+        definition = fmap (List.reverse .> List.head) dictionary
     in
         testCase "Should `clone`"
-        $ definition `rmap` Shikensu.localPath >>= assertEq "cloned.md"
+        $ assertDef definition Shikensu.localPath "cloned.md"
 
 
 
@@ -83,10 +82,10 @@ testExclude :: TestTree
 testExclude =
     let
         dictionary = fmap (Contrib.exclude "example.md") example_md
-        length_ = fmap length dictionary
+        dictLength = fmap length dictionary
     in
         testCase "Should `exclude`"
-        $ length_ >>= assertEq 0
+        $ assertEqm dictLength 0
 
 
 
@@ -108,33 +107,35 @@ testMetadata =
         -- 1. Insert C
         -- 2. Replace with A
         -- 3. Insert B
-        dictionary = example_md
-            <&> ( Contrib.insertMetadata (HashMap.fromList [ (keyC, valueC) ])
-               .> Contrib.replaceMetadata (HashMap.fromList [ (keyA, valueA) ])
-               .> Contrib.copyPropsToMetadata
-               .> Contrib.insertMetadata (HashMap.fromList [ (keyB, valueB) ])
+        dictionary = fmap
+            ( id
+                .> Contrib.insertMetadata   (HashMap.fromList [ (keyC, valueC) ])
+                .> Contrib.replaceMetadata  (HashMap.fromList [ (keyA, valueA) ])
+                .> Contrib.copyPropsToMetadata
+                .> Contrib.insertMetadata   (HashMap.fromList [ (keyB, valueB) ])
             )
+            example_md
 
-        definition = fmap (List.head . List.reverse) dictionary
+        definition      = fmap (List.reverse .> List.head) dictionary
 
-        lookupTitle = \def -> HashMap.lookup keyA (Shikensu.metadata def)
-        lookupHello = \def -> HashMap.lookup keyB (Shikensu.metadata def)
-        lookupRemoved = \def -> HashMap.lookup keyC (Shikensu.metadata def)
-        lookupBasename = \def -> HashMap.lookup keyBase (Shikensu.metadata def)
+        lookupTitle     = \def -> HashMap.lookup keyA (Shikensu.metadata def)
+        lookupHello     = \def -> HashMap.lookup keyB (Shikensu.metadata def)
+        lookupRemoved   = \def -> HashMap.lookup keyC (Shikensu.metadata def)
+        lookupBasename  = \def -> HashMap.lookup keyBase (Shikensu.metadata def)
     in
         testGroup
             "Metadata"
             [ testCase "Should no longer have `removed` key"
-            $ definition `rmap` lookupRemoved >>= assertEq Nothing
+            $ assertDef definition lookupRemoved Nothing
 
             , testCase "Should have `hello` key"
-            $ definition `rmap` lookupHello >>= assertEq (Just valueB)
+            $ assertDef definition lookupHello (Just valueB)
 
             , testCase "Should have `title` key"
-            $ definition `rmap` lookupTitle >>= assertEq (Just valueA)
+            $ assertDef definition lookupTitle (Just valueA)
 
             , testCase "Should have `basename` key"
-            $ definition `rmap` lookupBasename >>= assertEq (Just valueBase)
+            $ assertDef definition lookupBasename (Just valueBase)
             ]
 
 
@@ -143,18 +144,18 @@ testPermalink :: TestTree
 testPermalink =
     let
         dictionary = fmap (Contrib.permalink "index") example_md
-        definition = fmap (List.head) dictionary
+        definition = fmap List.head dictionary
     in
         testGroup
             "Permalink"
             [ testCase "Should have the correct `localPath`"
-            $ definition `rmap` Shikensu.localPath >>= assertEq "example/index.md"
+            $ assertDef definition Shikensu.localPath "example/index.md"
 
             , testCase "Should have the correct `parentPath`"
-            $ definition `rmap` Shikensu.parentPath >>= assertEq (Just "../")
+            $ assertDef definition Shikensu.parentPath (Just "../")
 
             , testCase "Should have the correct `pathToRoot`"
-            $ definition `rmap` Shikensu.pathToRoot >>= assertEq "../"
+            $ assertDef definition Shikensu.pathToRoot "../"
             ]
 
 
@@ -168,13 +169,13 @@ testPrefixDirname =
         testGroup
             "PrefixDirname"
             [ testCase "Should have the correct `dirname`"
-            $ definition `rmap` Shikensu.dirname >>= assertEq "prefix/fixtures"
+            $ assertDef definition Shikensu.dirname "prefix/fixtures"
 
             , testCase "Should have the correct `pathToRoot`"
-            $ definition `rmap` Shikensu.pathToRoot >>= assertEq "../../"
+            $ assertDef definition Shikensu.pathToRoot "../../"
 
             , testCase "Should have the correct `parentPath`"
-            $ definition `rmap` Shikensu.parentPath >>= assertEq (Just "../")
+            $ assertDef definition Shikensu.parentPath (Just "../")
             ]
 
 
@@ -182,14 +183,11 @@ testPrefixDirname =
 testRead :: TestTree
 testRead =
     let
-        dictionary = example_md >>= Contrib.IO.read
-        definition = fmap List.head dictionary
+        theResult  = Just (Text.pack "# Example\n")
+        definition = fmap List.head (example_md >>= Contrib.IO.read)
     in
         testCase "Should `read`"
-        $ definition
-            <&> Shikensu.content
-            <&> fmap Text.decodeUtf8
-            >>= assertEq (Just (Text.pack "# Example\n"))
+        $ assertDef definition (Shikensu.content .> fmap Text.decodeUtf8) theResult
 
 
 
@@ -197,10 +195,10 @@ testRename :: TestTree
 testRename =
     let
         dictionary = fmap (Contrib.rename "example.md" "renamed.md") example_md
-        definition = fmap (List.head) dictionary
+        definition = fmap List.head dictionary
     in
         testCase "Should `rename`"
-        $ definition `rmap` Shikensu.localPath >>= assertEq "renamed.md"
+        $ assertDef definition Shikensu.localPath "renamed.md"
 
 
 
@@ -208,25 +206,22 @@ testRenameExt :: TestTree
 testRenameExt =
     let
         dictionary = fmap (Contrib.renameExt ".md" ".html") example_md
-        definition = fmap (List.head) dictionary
+        definition = fmap List.head dictionary
     in
         testCase "Should `renameExt`"
-        $ definition `rmap` Shikensu.extname >>= assertEq ".html"
+        $ assertDef definition Shikensu.extname ".html"
 
 
 
 testRenderContent :: TestTree
 testRenderContent =
     let
+        theResult  = Just (Text.pack "<html># Example\n</html>")
         dictionary = fmap (Contrib.renderContent renderer) (example_md >>= Contrib.IO.read)
-        definition = fmap (List.head) dictionary
-        expectedResult = Just (Text.pack "<html># Example\n</html>")
+        definition = fmap List.head dictionary
     in
         testCase "Should `renderContent`"
-        $ definition
-            <&> Shikensu.content
-            <&> fmap Text.decodeUtf8
-            >>= assertEq expectedResult
+        $ assertDef definition (Shikensu.content .> fmap Text.decodeUtf8) theResult
 
 
 
@@ -234,11 +229,13 @@ testWrite :: TestTree
 testWrite =
     let
         destination = "tests/build/"
-        dictionary = list "tests/**/example.md" >>= Contrib.IO.read >>= Contrib.IO.write destination
+        dictionary  = list "tests/**/example.md"
+            >>= Contrib.IO.read
+            >>= Contrib.IO.write destination
+
         definition = fmap List.head dictionary
     in
         testCase "Should `write`"
-        $ definition
-            <&> Shikensu.rootDirname
-            >>= \r -> Text.readFile (joinPath [r, destination, "fixtures/example.md"])
-            >>= \c -> assertEq "# Example\n" (Text.unpack c)
+        $ fmap Shikensu.rootDirname definition
+            >>= \r -> readFile (joinPath [r, destination, "fixtures/example.md"])
+            >>= \c -> assertEq "# Example\n" c