+++ tests/Test/Utilities.hs
@@ -3,7 +3,7 @@ module Test.Utilities
     ) where
 
 import Flow
-import Shikensu.Types
+import Shikensu
 import Shikensu.Utilities as Utils
 import Test.Helpers
 import Test.Tasty
@@ -12,14 +12,13 @@ import Test.Tasty.HUnit
 import qualified Data.HashMap.Strict as HashMap (singleton)
 import qualified Data.List as List (head)
 import qualified Data.Tuple as Tuple (fst)
-import qualified Shikensu
 import qualified Shikensu.Contrib as Contrib
 
 
 utilityTests :: TestTree
 utilityTests = testGroup
     "Utility tests"
-    [testSequencing, testThunder]
+    [ testSequencing, testMetadataAccessors ]
 
 
 
@@ -31,40 +30,33 @@ testSequencing =
     let
         result =
             Utils.lsequence
-                [ ( "a", rootPath >>= Shikensu.list ["tests/fixtures/example.md"] )
-                , ( "b", rootPath >>= Shikensu.list ["tests/fixtures/example.md"] )
+                [ ( "a", Shikensu.listRelative ["tests/fixtures/example.md"] "." )
+                , ( "b", Shikensu.listRelative ["tests/fixtures/example.md"] "." )
                 ]
     in
         testCase "Test lsequence"
-        $ (List.head .> Tuple.fst) <$> result >>= assertEq "a"
+        $ fmap (List.head .> Tuple.fst) result >>= assertEq "a"
 
 
-testThunder :: TestTree
-testThunder =
+testMetadataAccessors :: TestTree
+testMetadataAccessors =
     let
-        flow =
-            Contrib.insertMetadata (HashMap.singleton "a" "Hi!")
-            .> return
-
-        dictionary =
-            rootPath
-                >>= Shikensu.list ["tests/fixtures/example.md"]
-                >>= flow
-
-        def =
-            fmap List.head dictionary
+        definition =
+            "."
+                |> define "tests/fixtures/example.md"
+                |> fmap (Contrib.insertMetadataDef $ HashMap.singleton "a" "Hi!")
 
         resultExisting =
-            fmap (\d -> metadata d ~> "a" :: Maybe String) def
+            fmap (\d -> metadata d ~> "a" :: Maybe String) definition
 
         resultNonExisting =
-            fmap (\d -> metadata d ~> "b" :: Maybe String) def
+            fmap (\d -> metadata d ~> "b" :: Maybe String) definition
     in
         testGroup
             "Test (~>)"
             [ testCase "Existing"
-            $ resultExisting >>= assertEq (Just "Hi!")
+            $ assertEqm resultExisting (Just "Hi!")
 
             , testCase "Non-existing"
-            $ resultNonExisting >>= assertEq Nothing
+            $ assertEqm resultNonExisting Nothing
             ]